- refactor index.js & project.js into main.js & projectDom.js
