+++ src/taskDom.js
@@ -1,96 +1,22 @@
-class Task {
-  constructor(title, description, dueDate, priority, project = 'default') {
-    this.title = title;
-    this.description = description;
-    this.dueDate = dueDate;
-    this.priority = priority;
-    this.project = project;
-    this.exists = true;
-  }
-
-  // Edit Delete Tasks methods
-  static deleteTask(projectTasks, taskProject, i, TASKS_LIST, renderTask) {
-    TASKS_LIST.addEventListener('click', (e) => {
-      if (e.target.id === `delete-btn-${i}`) {
-        projectTasks[i].exists = false;
-        renderTask(taskProject, projectTasks);
-      }
-    });
-  }
-
-  static editTask(projectTasks, taskProject, i, TASKS_LIST, renderTask) {
-    TASKS_LIST.addEventListener('click', (e) => {
-      if (e.target.id === `edit-btn-${i}`) {
-        const EDIT_NAME = document.querySelector(`#edit-task-name-${i}`).value;
-        const EDIT_DESC = document.querySelector(`#edit-task-description-${i}`).value;
-        const EDIT_DATE = document.querySelector(`#edit-task-date-${i}`).value;
-        const EDIT_PRIOR = document.querySelector(`#edit-task-priority-${i}`).value;
-
-        if (EDIT_NAME === '') {
-          return;
-        }
-        if (EDIT_DESC === '') {
-          return;
-        }
-        if (EDIT_DATE === '') {
-          return;
-        }
-        if (EDIT_PRIOR === '') {
-          return;
-        }
-
-        projectTasks[i].title = EDIT_NAME;
-        projectTasks[i].description = EDIT_DESC;
-        projectTasks[i].dueDate = EDIT_DATE;
-        projectTasks[i].priority = EDIT_PRIOR;
-
-        renderTask(taskProject, projectTasks);
-      }
-    });
-  }
-}
-
-// Tasks Form Information
-const TASKS_FORM = document.forms['tasks-form'];
-
-// prevent form default behaviour
-TASKS_FORM.addEventListener('submit', (e) => {
-  e.preventDefault();
-});
-
 // Task Creation Module (IIFE);
 export const taskCreation = (() => {
+  // Task Form
+  const taskForm = () => {
+    // Tasks Form Information
+    const TASKS_FORM = document.forms['tasks-form'];
+
+// prevent form default behaviour
+    TASKS_FORM.addEventListener('submit', (e) => {
+      e.preventDefault();
+    });
+
+    return TASKS_FORM;
+  }
+
   // create task object
   const newTask = (tasksArr, name, description, date, priority, project) => {
     const newTask = new Task(name, description, date, priority, project);
     tasksArr.push(newTask);
-  };
-
-  // UI color logic
-  const getColor = (arr, i) => {
-    let color;
-    if (arr[i].priority === 'low') {
-      color = 'success';
-    } else if (arr[i].priority === 'medium') {
-      color = 'info';
-    } else {
-      color = 'danger';
-    }
-    return color;
-  };
-
-  const buttonColor = (color) => {
-    let btnEdit;
-    let btnDelete;
-    if (color === 'success' || color === 'info') {
-      btnEdit = 'warning';
-      btnDelete = 'danger';
-    }
-    if (color === 'danger') {
-      btnEdit = 'primary';
-      btnDelete = 'warning';
-    }
-    return [btnEdit, btnDelete];
   };
 
   // render tasks in the UI
@@ -190,7 +116,8 @@ export const taskCreation = (() => {
   };
 
   // create task and display it on the UI
-  const createNewTask = (tasksArr) => {
+  const createNewTask = (tasksArr, taskForm) => {
+    const TASKS_FORM = taskForm();
     const TASK_BTN = document.querySelector('#task-btn');
     TASK_BTN.addEventListener('click', () => {
       const TASK_NAME = TASKS_FORM.querySelector('#task-name').value;
@@ -225,7 +152,7 @@ export const taskCreation = (() => {
       // render task to the UI
       renderTask(TASK_PROJECT, tasksArr);
 
-      document.querySelector('#tasks-form');
+      document.querySelector('#tasks-form').reset();
     });
   };
